rename: footer components divide
